Implementing steps to aggregate results form exposure windows

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/risk/result/AggregatedRiskResult.kt
@@ -2,11 +2,10 @@ package de.rki.coronawarnapp.risk.result
 
 import de.rki.coronawarnapp.server.protocols.internal.v2.RiskCalculationParametersOuterClass
 
-// TODO("Adjust Types")
 data class AggregatedRiskResult(
     val totalRiskLevel: RiskCalculationParametersOuterClass.NormalizedTimeToRiskLevelMapping.RiskLevel,
     val totalMinimumDistinctEncountersWithLowRisk: Int,
     val totalMinimumDistinctEncountersWithHighRisk: Int,
-    val mostRecentDateWithLowRisk: Int?,
-    val mostRecentDateWithHighRisk: Int?
+    val mostRecentDateWithLowRisk: Long?,
+    val mostRecentDateWithHighRisk: Long?
 )
